Renamed logerHandler to LoggerHandler. Refactored the code to use the new name.
