+++ project_code/helpers/DateNormalizer.py
@@ -5,7 +5,7 @@ from project_code.helpers.Preprocessing import PreprocessingDates
 from datetime import datetime, timedelta
 import calendar
 import re
-from project_code.helpers.logerHandler import setup_logger
+from project_code.helpers.LogerHandler import setup_logger
 
 class DatesExplorer:
     """A class for exploring and validating date columns in a DataFrame."""
